+++ data/cars.js
@@ -1,0 +1,26 @@
+const cars =  [
+    {
+        name: 'car 1',
+        shortInfo: 'electric one, you will enjoy with it',
+        description: 'this is the most comfortable car in our garage, you can try it before rent'
+    },{
+        name: 'car 2',
+        shortInfo: 'electric one, you will enjoy with it',
+        description: 'this is the most comfortable car in our garage, you can try it before rent'
+    },{
+        name: 'car 3',
+        shortInfo: 'electric one, you will enjoy with it',
+        description: 'this is the most comfortable car in our garage, you can try it before rent'
+    },{
+        name: 'car 4',
+        shortInfo: 'electric one, you will enjoy with it',
+        description: 'this is the most comfortable car in our garage, you can try it before rent'
+    },{
+        name: 'car 5',
+        shortInfo: 'electric one, you will enjoy with it',
+        description: 'this is the most comfortable car in our garage, you can try it before rent'
+    },
+]
+
+
+export default cars
