+++ index.html
@@ -17,7 +17,7 @@
     ></script>
   </head>
   <body>
-    <img class="home-bg d-none" src="./assets/background.jpg" alt="carbg" />
+    <img class="home-bg" src="./assets/background.jpg" alt="carbg" />
     <nav id="toolbar">
       <a class="nav-bars">
         <i class="fa-sharp fa-solid fa-bars fs-1 fw-bold"></i>
@@ -40,7 +40,11 @@
       </p>
     </section>
     <section id="services" class="services d-flex flex-column align-items-center">
-      <h3 class="section-title text-light">Main Services</h3>
+      <h3 class="section-title text-light mt-4">Main Services</h3>
+      <hr class="bord"></hr>
+    </section>
+    <section class="cars d-flex flex-column align-items-center mt-4 min-vh-100">
+      <h3 class="section-title">Our Cars</h3>
       <hr class="bord"></hr>
     </section>
     <script type="module" src="./script.js"></script>
