+++ script.js
@@ -1,5 +1,5 @@
 import generateServiceCard from './utils/serviceCard.js'
+import generateCarCard from './utils/carCard.js'
 
 generateServiceCard()
-
-
+generateCarCard()
