--- services.js
@@ -1,30 +1,0 @@
-const services = [
-  {
-    logo: "../assets/icons/car-seat.png",
-    serviceTitle: "Comfort",
-    serviceDescription: "travel anywhere with high comfort, less amount of money and friendly driver",
-  },
-  {
-    logo: "../assets/icons/fast.png",
-    serviceTitle: "Fast",
-    serviceDescription: "travel anywhere with high comfort, less amount of money and friendly driver",
-  },
-  {
-    logo: "../assets/icons/time.png",
-    serviceTitle: "Time Dedication",
-    serviceDescription: "travel anywhere with high comfort, less amount of money and friendly driver",
-  },
-  {
-    logo: "../assets/icons/friendly.png",
-    serviceTitle: "friendly",
-    serviceDescription: "travel anywhere with high comfort, less amount of money and friendly driver",
-  },
-  {
-    logo: "../assets/icons/tech.png",
-    serviceTitle: "technology",
-    serviceDescription: "travel anywhere with high comfort, less amount of money and friendly driver",
-  },
-];
-
-
-export default services;
