+++ utils/carCard.js
@@ -1,0 +1,7 @@
+import cars from "../data/cars.js";
+
+const generateCarCard = () => {
+    
+};
+
+export default generateCarCard;
